Fixed run.sh options, and persisting cache and key by default
The SSL certificate directory was erroneously being set to use
the cache directory ($CACHEDIR).

Also, the default directories were in /tmp. This meant that every
time the host computer was restarted, the cache - and more
importantly, the CA certificate - was destroyed. Images that had
been built using the old CA certificate would no longer be able to
use the proxy, so it's important that the CA certificate at least
is persisted.